requirements.txt
anyio==4.3.0
APScheduler==3.6.3
beautifulsoup4==4.12.3
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
dadata==21.10.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tornado==6.1
tzlocal==5.2
urllib3==2.2.1
